move to-otel-attributes
