+++ packages/ai/src/otel/utils/to-otel-attribute.ts
