+++ packages/ai/test/otel/utils/to-otel-attribute.test.ts
@@ -1,5 +1,5 @@
+import { toOtelAttribute } from 'src/otel/utils/to-otel-attribute';
 import { describe, it, expect } from 'vitest';
-import { toOtelAttribute } from '../../src/util/to-otel-attribute.js';
 
 describe('toOtelAttribute', () => {
   it('handles string', () => {
@@ -29,30 +29,25 @@ describe('toOtelAttribute', () => {
     expect(toOtelAttribute(Symbol('test'))).toBeUndefined();
   });
 
-  // Date
   it('handles date', () => {
     const date = new Date('2023-01-01T00:00:00.000Z');
     expect(toOtelAttribute(date)).toBe('2023-01-01T00:00:00.000Z');
   });
 
-  // Object
   it('handles object', () => {
     expect(toOtelAttribute({ key: 'value' })).toBe('{"key":"value"}');
   });
 
-  // Nested object
   it('handles nested object', () => {
     const obj = { a: 1, b: { c: 'nested' } };
     expect(toOtelAttribute(obj)).toBe('{"a":1,"b":{"c":"nested"}}');
   });
 
-  // Array of primitives
   it('handles array of primitives', () => {
     expect(toOtelAttribute([1, 'two', true])).toEqual([1, 'two', true]);
     expect(toOtelAttribute([undefined, null, NaN])).toBeUndefined(); // all filtered out
   });
 
-  // Array of objects
   it('handles array of objects', () => {
     const arr = [{ a: 1 }, new Date('2023-01-01')];
     expect(toOtelAttribute(arr)).toEqual(['{"a":1}', '2023-01-01T00:00:00.000Z']);
